other/requirements.txt
numpy>=1.26.0
torch>=2.1.0
gymnasium>=0.29.0
mujoco>=3.0.0
matplotlib>=3.8.0
tensorboard>=2.15.0
pyyaml>=6.0.1
tqdm>=4.62.0
stable-baselines3>=2.0.0 